export the feature functions separately for use outside of hooks
